fix bundling issue with Next.js 13
