+++ packages/nextjs/src/server/utils.ts
@@ -21,30 +21,8 @@ import {
   missingSignInUrlInDev,
 } from './errors';
 import { errorThrower } from './errorThrower';
+import { detectClerkMiddleware, isNextRequest } from './headers-utils';
 import type { RequestLike } from './types';
-
-export function getCustomAttributeFromRequest(req: RequestLike, key: string): string | null | undefined {
-  // @ts-expect-error - TS doesn't like indexing into RequestLike
-  return key in req ? req[key] : undefined;
-}
-
-export function getAuthKeyFromRequest(
-  req: RequestLike,
-  key: keyof typeof constants.Attributes,
-): string | null | undefined {
-  return getCustomAttributeFromRequest(req, constants.Attributes[key]) || getHeader(req, constants.Headers[key]);
-}
-
-export function getHeader(req: RequestLike, name: string): string | null | undefined {
-  if (isNextRequest(req)) {
-    return req.headers.get(name);
-  }
-
-  // If no header has been determined for IncomingMessage case, check if available within private `socket` headers
-  // When deployed to vercel, req.headers for API routes is a `IncomingHttpHeaders` key-val object which does not follow
-  // the Headers spec so the name is no longer case-insensitive.
-  return req.headers[name] || req.headers[name.toLowerCase()] || (req.socket as any)?._httpMessage?.getHeader(name);
-}
 
 export function getCookie(req: RequestLike, name: string): string | undefined {
   if (isNextRequest(req)) {
@@ -59,19 +37,6 @@ export function getCookie(req: RequestLike, name: string): string | undefined {
     return typeof reqCookieOrString === 'string' ? reqCookieOrString : reqCookieOrString.value;
   }
   return req.cookies[name];
-}
-
-function isNextRequest(val: unknown): val is NextRequest {
-  try {
-    const { headers, nextUrl, cookies } = (val || {}) as NextRequest;
-    return (
-      typeof headers?.get === 'function' &&
-      typeof nextUrl?.searchParams.get === 'function' &&
-      typeof cookies?.get === 'function'
-    );
-  } catch (e) {
-    return false;
-  }
 }
 
 const OVERRIDE_HEADERS = 'x-middleware-override-headers';
@@ -196,10 +161,6 @@ export const handleMultiDomainAndProxy = (clerkRequest: ClerkRequest, opts: Auth
 export const redirectAdapter = (url: string | URL) => {
   return NextResponse.redirect(url, { headers: { [constants.Headers.ClerkRedirectTo]: 'true' } });
 };
-
-export function detectClerkMiddleware(req: RequestLike): boolean {
-  return Boolean(getAuthKeyFromRequest(req, 'AuthStatus'));
-}
 
 export function assertAuthStatus(req: RequestLike, error: string) {
   if (!detectClerkMiddleware(req)) {
